fix Bug #72220, 1. change DatabaseAuthenticationCacheService to interface, and add DatabaseAuthenticationCacheServiceImp, since  java Dynamic proxy requires interface. 2. Maintain the service as a singleton with atomic visibility 3. init service before connecting.
